view movie edits

+++ RealDeal/src/main/java/com/nashss/se/realdeal/models/ReviewsModel.java
@@ -1,6 +1,5 @@
 package com.nashss.se.realdeal.models;
 
-import java.time.LocalDate;
 import java.util.Objects;
 
 public class ReviewsModel {
@@ -8,7 +7,7 @@ public class ReviewsModel {
     private final String movieId;
     private final String username;
     private final String text;
-    private final int rating;
+    private final double rating;
     private final String movieDate;
 
     /**
@@ -22,7 +21,7 @@ public class ReviewsModel {
      * @param movieDate the date
      */
 
-    public ReviewsModel(String id, String movieId, String username, String text, int rating, String movieDate) {
+    public ReviewsModel(String id, String movieId, String username, String text, double rating, String movieDate) {
         this.id = id;
         this.movieId = movieId;
         this.username = username;
@@ -49,7 +48,7 @@ public class ReviewsModel {
         return text;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
@@ -75,7 +74,7 @@ public class ReviewsModel {
         private String movieId;
         private String username;
         private String text;
-        private int rating;
+        private double rating;
         private String movieDate;
 
         public Builder withId(String id) {
@@ -98,7 +97,7 @@ public class ReviewsModel {
             return this;
         }
 
-        public Builder withRating(int rating) {
+        public Builder withRating(double rating) {
             this.rating = rating;
             return this;
         }
